+++ packages/navbar/src/Navbar.tsx
@@ -7,6 +7,11 @@ import { RightSidebar } from './RightSidebar';
 import { RightPanelSection } from './RightPanelSection';
 import { Application, Broker } from './types';
 import { Wrapper } from './index.style';
+import { AnalyticsContextProvider } from '@lendi-ui/utils';
+import { analysticHelper } from './helpers/helpers';
+import { WindowPosition } from '@lendi/lendi-analytics-package';
+import { withTheme } from 'styled-components';
+import { ThemeMap } from '@lendi-ui/theme';
 
 export interface NavProps {
   isAuthenticated?: boolean;
@@ -15,12 +20,12 @@ export interface NavProps {
   onChat?: () => void;
   onLogout?: () => void;
   variant?: 'transparent' | 'white';
+  theme?: ThemeMap;
 }
 interface NavState {
   isLeftSidebarVisible?: boolean;
   isRightSidebarVisible?: boolean;
 }
-
 class Navbar extends React.Component<NavProps, NavState> {
   state = {
     isLeftSidebarVisible: false,
@@ -53,56 +58,63 @@ class Navbar extends React.Component<NavProps, NavState> {
 
   render() {
     const { isLeftSidebarVisible, isRightSidebarVisible } = this.state;
-    const { isAuthenticated, application, broker, onChat, onLogout, variant } = this.props;
-
+    const { isAuthenticated, application, broker, onChat, onLogout, variant, theme } = this.props;
     return (
       <Wrapper>
-        <Header
-          onOpenLeftSidebar={this.onOpenLeftSidebar}
-          onOpenRightSidebar={this.onOpenRightSidebar}
-          isAuthenticated={isAuthenticated}
-          continueApplicationUrl={application && application.continueURL}
-          isTransparent={!variant || variant !== 'white'}
-        />
-        <LeftSidebar
-          onLogout={onLogout}
-          show={isLeftSidebarVisible}
-          onHide={this.onCloseLeftSidebar}
-          isAuthenticated={isAuthenticated}
-          applicationDetails={
-            application && {
-              applicationNumber: application.number,
-              applicationStage: application.stage,
-            }
-          }
+        <AnalyticsContextProvider
+          value={{
+            analyticsForNavigation: (text: string, position: WindowPosition) =>
+              analysticHelper(text, !!broker, position, theme),
+          }}
         >
-          <LeftPanelSection
+          <Header
+            onOpenLeftSidebar={this.onOpenLeftSidebar}
+            onOpenRightSidebar={this.onOpenRightSidebar}
             isAuthenticated={isAuthenticated}
-            application={application}
-            applicants={application && application.applicants}
+            continueApplicationUrl={application && application.continueURL}
+            isTransparent={!variant || variant !== 'white'}
           />
-        </LeftSidebar>
-        <RightSidebar
-          onChat={onChat}
-          show={isRightSidebarVisible}
-          onHide={this.onCloseRightSidebar}
-          broker={
-            broker && {
-              fullName: broker.fullName || '',
-              photo: broker.photo || '',
-              title: broker.title || '',
-            }
-          }
-        >
-          <RightPanelSection
+          <LeftSidebar
             onChat={onChat}
-            applicationNumber={application && application.number}
-            phoneNumber={broker && broker.phone}
-          />
-        </RightSidebar>
+            onLogout={onLogout}
+            show={isLeftSidebarVisible}
+            onHide={this.onCloseLeftSidebar}
+            isAuthenticated={isAuthenticated}
+            applicationDetails={
+              application && {
+                applicationNumber: application.number,
+                applicationStage: application.stage,
+              }
+            }
+          >
+            <LeftPanelSection
+              isAuthenticated={isAuthenticated}
+              application={application}
+              applicants={application && application.applicants}
+            />
+          </LeftSidebar>
+          <RightSidebar
+            onChat={onChat}
+            show={isRightSidebarVisible}
+            onHide={this.onCloseRightSidebar}
+            broker={
+              broker && {
+                fullName: broker.fullName || '',
+                photo: broker.photo || '',
+                title: broker.title || '',
+              }
+            }
+          >
+            <RightPanelSection
+              onChat={onChat}
+              applicationNumber={application && application.number}
+              phoneNumber={broker && broker.phone}
+            />
+          </RightSidebar>
+        </AnalyticsContextProvider>
       </Wrapper>
     );
   }
 }
 
-export default Navbar;
+export default withTheme(Navbar);