+++ packages/navbar/src/RightPanelSection/index.tsx
@@ -7,18 +7,19 @@ export interface RightPanelSectionProps {
   applicationNumber?: number;
   phoneNumber?: string;
   onChat?: () => void;
+  analyticsCallback?: (text: string) => void;
 }
 
 const toTel = (phoneNumber: string) => `tel:+61${phoneNumber.replace(/ /g, '')}`;
 
 export const RightPanelSection = (props: RightPanelSectionProps) => {
-  const { phoneNumber = PHONE, applicationNumber = 0, onChat } = props;
+  const { phoneNumber = PHONE, applicationNumber = 0 } = props;
   return (
     <RightSidebar.Section>
       <RightSidebar.Item icon={<Telephone color="primary.500" />} href={phoneNumber && toTel(phoneNumber)}>
         {phoneNumber}
       </RightSidebar.Item>
-      <RightSidebar.Item icon={<LiveChat color="primary.500" />} onClick={onChat} labelText={'chat online'}>
+      <RightSidebar.Item icon={<LiveChat color="primary.500" />} labelText={'chat online'}>
         Chat online
       </RightSidebar.Item>
       <RightSidebar.Item icon={<PickATime color="primary.500" />} href="/appointment-booking">