+++ packages/navbar/src/Sidebar/index.tsx
@@ -4,6 +4,8 @@ import Overlay from '@lendi-ui/overlay';
 import Transition, { State } from '@lendi-ui/transition';
 import { Wrapper, CloseButton, OverlayWrapper } from './index.style';
 import { Close } from '@lendi-ui/icon';
+import { AnalyticsContext } from '@lendi-ui/utils';
+import { WindowPosition } from '@lendi/lendi-analytics-package';
 
 export interface SidebarProps {
   side: 'left' | 'right';
@@ -13,8 +15,9 @@ export interface SidebarProps {
 }
 
 class Sidebar extends React.Component<SidebarProps> {
+  static contextType: any = AnalyticsContext;
   render() {
-    const { side, show, onHide, children } = this.props;
+    const { side, show, onHide = () => {}, children } = this.props;
     return (
       <Fragment>
         <OverlayWrapper>
@@ -23,7 +26,13 @@ class Sidebar extends React.Component<SidebarProps> {
         <Transition active={show} timeout={250}>
           {(state: State) => (
             <Wrapper side={side} transition={state}>
-              <CloseButton onClick={onHide} aria-label="close">
+              <CloseButton
+                onClick={() => {
+                  this.context.analyticsForNavigation('icon', WindowPosition.navigation_left);
+                  onHide();
+                }}
+                aria-label="close"
+              >
                 <Close color="shade.300" />
               </CloseButton>
               {children}